envrequirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.8.2
blinker==1.6.3
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
decorator==5.1.1
duckdb==0.8.1
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.0
openai==0.27.10
oscrypto==1.3.0
packaging==23.2
pandas==1.5.3
pandasai==1.3.2
parso==0.8.3
pdfminer.six==20221105
pdfreader==0.1.12
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
protobuf==4.24.4
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
scipy==1.11.3
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.2.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.49
stack-data==0.6.3
streamlit==1.27.2
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.17
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
yarl==1.9.2
zipp==3.17.0
